refactor(query-audit): pg.rs → postgres.rs (review: no abbreviated file names)
sqlx's own convention, adopted wholesale: the MODULE is `postgres`, only
type names carry the `Pg` prefix (sqlx::postgres::PgPool; here
postgres::PgAudit). No behaviour change.
